+++ task/default/styles.js
@@ -18,7 +18,7 @@ gulp.task('styles:common', function() {
 		.pipe($.include())
 		.pipe($.sass())
 		.pipe($.autoprefixer())
-		.pipe($.if($.util.env.production, $.csso()))
+		.pipe($.if($.util.env.production, $.cssnano()))
 		.pipe($.sourcemaps.write('.'))
 		.pipe(gulp.dest(path.public.style))
 		.pipe($.duration(name))