+++ source/script/vendor.js
@@ -2,5 +2,6 @@
 
 //= include ../../bower_components/modernizr/modernizr.js
 //= include ../../bower_components/metaquery/metaquery.js
+//= include ../../bower_components/picturefill/dist/picturefill.js
 //= include ../../bower_components/velocity/velocity.js
 //= include ../../bower_components/velocity/velocity.ui.js