+++ source/script/vendor.js
@@ -1,7 +1,7 @@
 // vendor.js
 
-//= include ../../bower_components/modernizr/modernizr.js
-//= include ../../bower_components/metaquery/metaquery.js
-//= include ../../bower_components/picturefill/dist/picturefill.js
-//= include ../../bower_components/velocity/velocity.js
-//= include ../../bower_components/velocity/velocity.ui.js
+// = include ../../bower_components/modernizr/modernizr.js
+// = include ../../bower_components/metaquery/metaquery.js
+// = include ../../bower_components/picturefill/dist/picturefill.js
+// = include ../../bower_components/velocity/velocity.js
+// = include ../../bower_components/velocity/velocity.ui.js