+++ task/default/scripts.js
@@ -2,11 +2,23 @@
 
 var gulp = require('gulp'),
 	$ = require('gulp-load-plugins')(),
+	config = require('../config'),
 	helper = require('../helper'),
 	path = require('../path');
 
 // Base
-gulp.task('scripts', ['scripts:common', 'scripts:vendor']);
+gulp.task('scripts', ['scripts:lint', 'scripts:common', 'scripts:vendor']);
+
+// Lint
+gulp.task('scripts:lint', function() {
+	var name = 'Lint Scripts';
+
+	return gulp.src([path.source.script + '**/*.js', '!' + path.source.script + 'vendor/**'])
+		.pipe($.plumber(helper.error))
+		.pipe($.xo(config.plugin.xo))
+		.pipe($.duration(name))
+		.pipe(helper.success(name));
+});
 
 // Common
 gulp.task('scripts:common', function() {